backend/requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6


# Database
supabase>=2.25.1
psycopg2-binary==2.9.10

# Environment and Config
python-dotenv==1.0.0
pydantic>=2.11.7,<3.0
pydantic-settings>=2.3.0
email-validator>=2.0.0
# AI Models
openai==1.3.0
google-generativeai>=0.8.0
groq>=0.10.0

# Voice Services
elevenlabs==0.2.27
deepgram-sdk==3.2.7

# File Processing
PyPDF2==3.0.1
python-docx==1.1.0
pypandoc==1.12

# Utilities
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx>=0.26,<0.29  # For downloading files for email attachments (compatible with openai, deepgram, supabase)

# Logging and Monitoring
structlog==23.2.0

# Rate Limiting
slowapi==0.1.9

# Email Services
resend==2.1.0  # Resend email API
jinja2==3.1.2  # Template engine for email templates

# Calendar Integration
google-api-python-client>=2.111.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-env==1.1.3
httpx==0.26.0  # For async test client
faker==22.0.0  # For generating test data
freezegun==1.4.0  # For time mocking


